Switching simple_component to be component
 - Moving implementation of a simple component to component(_base)
 - Keeping using simple_component = component and friends for backwards
   compatbility.
