+++ hpx/lcos/base_lco_with_value.hpp
@@ -14,7 +14,6 @@
 #include <hpx/runtime/components/component_type.hpp>
 #include <hpx/runtime/components/server/component.hpp>
 #include <hpx/runtime/components/server/managed_component_base.hpp>
-#include <hpx/runtime/components/server/simple_component_base.hpp>
 #include <hpx/runtime/components_fwd.hpp>
 #include <hpx/runtime/naming/id_type.hpp>
 #include <hpx/traits/is_component.hpp>
@@ -44,12 +43,6 @@ namespace hpx { namespace lcos
         struct base_lco_wrapping_type<traits::detail::component_tag, BaseLco>
         {
             typedef components::component<BaseLco> type;
-        };
-
-        template <typename BaseLco>
-        struct base_lco_wrapping_type<traits::detail::simple_component_tag, BaseLco>
-        {
-            typedef components::simple_component<BaseLco> type;
         };
 
         template <typename BaseLco>
@@ -201,24 +194,7 @@ namespace hpx { namespace traits
     {
         static components::component_type get()
         {
-            return components::component_base_lco_with_value_unmanaged;
-        }
-
-        static void set(components::component_type)
-        {
-            HPX_ASSERT(false);
-        }
-    };
-
-    template <typename Result, typename RemoteResult, typename Enable>
-    struct component_type_database<
-        hpx::lcos::base_lco_with_value<
-            Result, RemoteResult, traits::detail::simple_component_tag
-        >, Enable>
-    {
-        static components::component_type get()
-        {
-            return components::component_base_lco_with_value_simple_unmanaged;
+            return components::component_base_lco_with_value;
         }
 
         static void set(components::component_type)
