+++ hpx/runtime/actions/set_lco_value_continuation.hpp
@@ -30,20 +30,6 @@ namespace hpx { namespace actions
     };
 
     ///////////////////////////////////////////////////////////////////////////
-    struct set_lco_value_simple_unmanaged_continuation
-    {
-        template <typename T>
-        HPX_FORCEINLINE T operator()(naming::id_type const& lco, T&& t) const
-        {
-            hpx::set_lco_value_simple_unmanaged(lco, std::forward<T>(t));
-
-            // Yep, 't' is a zombie, however we don't use the returned value
-            // anyways. We need it for result type calculation, though.
-            return std::move(t);
-        }
-    };
-
-    ///////////////////////////////////////////////////////////////////////////
     struct set_lco_value_unmanaged_continuation
     {
         template <typename T>
