+++ hpx/runtime/components/component_type.hpp
@@ -41,33 +41,30 @@ namespace hpx { namespace components
         component_base_lco = 4,
 
         // Base component for LCOs that produce values.
-        component_base_lco_with_value_simple_unmanaged = 5,
-
-        // Base component for LCOs that produce values.
-        component_base_lco_with_value_unmanaged = 6,
+        component_base_lco_with_value_unmanaged = 5,
 
         // (Managed) base component for LCOs that produce values.
-        component_base_lco_with_value = 7,
+        component_base_lco_with_value = 6,
 
         // Synchronization latch, barrier, and flex_barrier LCOs.
-        component_latch = ((8 << 16) | component_base_lco_with_value),
-        component_barrier = ((9 << 16) | component_base_lco),
-        component_flex_barrier = ((10 << 16) | component_base_lco),
+        component_latch = ((7 << 16) | component_base_lco_with_value),
+        component_barrier = ((8 << 16) | component_base_lco),
+        component_flex_barrier = ((9 << 16) | component_base_lco),
 
         // An LCO representing a value which may not have been computed yet.
-        component_promise = ((11 << 16) | component_base_lco_with_value),
+        component_promise = ((10 << 16) | component_base_lco_with_value),
 
         // AGAS locality services.
-        component_agas_locality_namespace = 12,
+        component_agas_locality_namespace = 11,
 
         // AGAS primary address resolution services.
-        component_agas_primary_namespace = 13,
+        component_agas_primary_namespace = 12,
 
         // AGAS global type system.
-        component_agas_component_namespace = 14,
+        component_agas_component_namespace = 13,
 
         // AGAS symbolic naming services.
-        component_agas_symbol_namespace = 15,
+        component_agas_symbol_namespace = 14,
 
         component_last,
         component_first_dynamic = component_last,
